+++ mu-loader/mu-loader.php
@@ -1,0 +1,11 @@
+<?php
+/**
+ * Autoloads the must-use plugins.
+ *
+ * @since       3.0.5
+ * @author      hellofromTonya
+ * @link        https://github.com/wpfulcrum/fulcrum
+ * @license     GNU-2.0+
+ */
+
+require __DIR__ . '/fulcrum/bootstrap.php';